SEQNG-468 Revamp of the logic to the set the url and actions
