+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/components/sequence/toolbars/StepConfigToolbar.scala
@@ -6,7 +6,7 @@ package seqexec.web.client.components.sequence.toolbars
 import cats.implicits._
 import gem.Observation
 import japgolly.scalajs.react.extra.router.RouterCtl
-import japgolly.scalajs.react.{Callback, ScalaComponent}
+import japgolly.scalajs.react.ScalaComponent
 import japgolly.scalajs.react.vdom.html_<^._
 import japgolly.scalajs.react.component.Scala.Unmounted
 import seqexec.model.enum.Instrument
@@ -14,7 +14,6 @@ import seqexec.web.client.model.Pages._
 import seqexec.web.client.model.RunningStep
 import seqexec.web.client.circuit.SeqexecCircuit
 import seqexec.web.client.components.SeqexecStyles
-import seqexec.web.client.actions.{SelectIdToDisplay, ShowStepConfig, SelectSequencePreview}
 import seqexec.web.client.semanticui.elements.button.ButtonGroup
 import seqexec.web.client.semanticui.elements.button.Button
 import seqexec.web.client.semanticui.elements.button.Button._
@@ -30,18 +29,6 @@ import mouse.boolean._
 object StepConfigToolbar {
   final case class Props(router: RouterCtl[SeqexecPages], instrument: Instrument, id: Observation.Id, step: Int, total: Int, isPreview: Boolean)
 
-  def backToSequence(p: Props): Callback =
-    SeqexecCircuit.dispatchCB(SelectSequencePreview(p.instrument, p.id, p.step)).when(p.isPreview) *>
-    SeqexecCircuit.dispatchCB(SelectIdToDisplay(p.instrument, p.id)).unless(p.isPreview) *>
-    Callback.empty
-
-  def previousStep(p: Props): Callback =
-    SeqexecCircuit.dispatchCB(ShowStepConfig(p.instrument, p.id, p.step))
-
-  def nextStep(p: Props): Callback =
-    SeqexecCircuit.dispatchCB(ShowStepConfig(p.instrument, p.id, p.step + 1))
-
-  @SuppressWarnings(Array("org.wartremover.warts.OptionPartial"))
   private val component = ScalaComponent.builder[Props]("StepConfigToolbar")
     .stateless
     .render_P { p =>
@@ -78,21 +65,24 @@ object StepConfigToolbar {
         <.div(
           ^.cls := "ui left floated eight wide column",
           SeqexecStyles.shorterFields,
+          // Back to sequence button
           p.router.link(sequencePage)
-            (Button(Button.Props(icon = Some(IconChevronLeft), labeled = LeftLabeled, onClick = backToSequence(p)), "Back"))),
+            (Button(Button.Props(icon = Some(IconChevronLeft), labeled = LeftLabeled, onClick = p.router.setUrlAndDispatchCB(sequencePage)), "Back"))),
           <.div(
             ^.cls := "ui right floated eight wide column",
             SeqexecStyles.shorterFields,
             ButtonGroup(
               ButtonGroup.Props(List(GStyle.fromString("right floated"))),
+              // Previous step button
               (p.step > 0).option(p.router.link(prevStepPage)
-                (Button(Button.Props(icon = Some(IconChevronLeft), labeled = LeftLabeled, onClick = previousStep(p)), "Prev"))),
+                (Button(Button.Props(icon = Some(IconChevronLeft), labeled = LeftLabeled, onClick = p.router.setUrlAndDispatchCB(prevStepPage)), "Prev"))),
               Label(Label.Props(
                 RunningStep(p.step, p.total).show,
                 size = Size.Large,
                 extraStyles = List(SeqexecStyles.labelAsButton))),
+              // Next step button
               (p.step < p.total - 1).option(p.router.link(nextStepPage)
-                (Button(Button.Props(icon = Some(IconChevronRight), labeled = RightLabeled, onClick = nextStep(p)), "Next")))
+                (Button(Button.Props(icon = Some(IconChevronRight), labeled = RightLabeled, onClick = p.router.setUrlAndDispatchCB(nextStepPage)), "Next")))
               )
             )
           )
