+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/model/actions.scala
@@ -35,11 +35,13 @@ object actions {
   case object Logout extends Action
 
   // Action to select a sequence for display
-  final case class SelectIdToDisplay(i: Instrument, id: Observation.Id) extends Action
-  final case class SelectSequenceConfig(id: Observation.Id, step: StepId) extends Action
+  final case class SelectIdToDisplay(i: Instrument, id: Observation.Id, step: StepId) extends Action
+  // final case class SelectSequenceConfig(id: Observation.Id, step: StepId) extends Action
   final case class SelectSequencePreview(i: Instrument, id: Observation.Id, step: StepId) extends Action
   case object SelectEmptyPreview extends Action
-  case object CleanSequences extends Action
+  case object SelectRoot extends Action
+  final case class ShowStepConfig(i: Instrument, id: Observation.Id, step: Int) extends Action
+  final case class ShowPreviewStepConfig(i: Instrument, id: Observation.Id, step: Int) extends Action
 
   // Actions related to executing sequences
   final case class RequestRun(s: Observation.Id) extends Action
@@ -68,9 +70,6 @@ object actions {
   final case class RunObsPauseFailed(s: Observation.Id) extends Action
   final case class RunObsResumeFailed(s: Observation.Id) extends Action
 
-  final case class ShowStepConfig(i: Instrument, id: Observation.Id, step: Int) extends Action
-  final case class ShowPreviewStepConfig(i: Instrument, id: Observation.Id, step: Int) extends Action
-  final case class HideStepConfig(i: Instrument) extends Action
   final case class RememberCompleted(s: SequenceView) extends Action
 
   final case class AppendToLog(l: ServerLogMessage) extends Action
@@ -100,6 +99,7 @@ object actions {
   final case class UpdateLoadedSequences(ids: List[Observation.Id]) extends Action
   final case class UpdateOnLoadUpdate(i: Instrument, id: Observation.Id, loaded: List[Observation.Id]) extends Action
   final case class LoadSequence(observer: Observer, i: Instrument, id: Observation.Id) extends Action
+  case object CleanSequences extends Action
 
   // Used for UI debugging
   final case class MarkStepAsRunning(s: Observation.Id, step: Int) extends Action
