+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/model/handlers/InitialSyncHandler.scala
@@ -27,7 +27,7 @@ class InitialSyncHandler[M](modelRW: ModelRW[M, InitialSyncFocus]) extends Actio
     case ServerMessage(s: SeqexecModelUpdate) if runningSequence(s).isDefined && value.firstLoad =>
       val running = runningSequence(s)
       running.fold(updated(value.copy(firstLoad = true))) { f =>
-        val effect = Effect(Future(SelectIdToDisplay(f.metadata.instrument, f.id)))
+        val effect = Effect(Future(SelectIdToDisplay(f.metadata.instrument, f.id, 0)))
         updated(value.copy(firstLoad = false), effect)
       }
 
@@ -36,24 +36,26 @@ class InitialSyncHandler[M](modelRW: ModelRW[M, InitialSyncFocus]) extends Actio
       // the page maybe not in sync with the tabs. Let's fix that
       val sids = s.view.queue.map(_.id)
       val effect = value.location match {
-        case SequencePage(i, id, _) if sids.contains(id)                 =>
+        case p @ SequencePage(_, id, _) if sids.contains(id)       =>
           // We need to effect to update the reference
-          Effect(Future(SelectIdToDisplay(i, id)))
+          Effect(Future(PageActionP.reverseGet(p)))
 
-        case PreviewPage(i, id, st) if sids.contains(id)                 =>
+        case p @ SequenceConfigPage(_, id, _) if sids.contains(id) =>
+          // We need to effect to update the reference
+          Effect(Future(PageActionP.reverseGet(p)))
+
+        case p @ PreviewPage(i, id, st) if sids.contains(id)       =>
           val isLoaded = s.view.loaded.values.toList.contains(id)
           // We need to effect to update the reference
           if (isLoaded) {
-            Effect(Future(SelectIdToDisplay(i, id)))
+            val action = SelectIdToDisplay(i, id, st)
+            val pageE = PageActionP.getOption(action).map(p => Effect(Future(NavigateTo(p)))).getOrElse(VoidEffect)
+            Effect(Future(action)) >> pageE
           } else {
-            Effect(Future(SelectSequencePreview(i, id, st)))
+            Effect(Future(PageActionP.reverseGet(p)))
           }
 
-        case SequenceConfigPage(i, id, st) if sids.contains(id)          =>
-          // We need to effect to update the reference
-          Effect(Future(ShowStepConfig(i, id, st)))
-
-        case PreviewConfigPage(i, id, st) if sids.contains(id)           =>
+        case PreviewConfigPage(i, id, st) if sids.contains(id)     =>
           val isLoaded = s.view.loaded.values.toList.contains(id)
           // We need to effect to update the reference
           if (isLoaded) {
@@ -62,7 +64,7 @@ class InitialSyncHandler[M](modelRW: ModelRW[M, InitialSyncFocus]) extends Actio
             Effect(Future(ShowPreviewStepConfig(i, id, st)))
           }
 
-        case _                                                           =>
+        case _                                                     =>
           // No matches
           VoidEffect
       }
