+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/model/handlers/NavigationHandler.scala
@@ -4,12 +4,9 @@
 package seqexec.web.client.handlers
 
 import cats.implicits._
-import diode.{ActionHandler, ActionResult, Effect, ModelRW}
+import diode.{ActionHandler, ActionResult, /*Effect,*/ ModelRW}
 import seqexec.web.client.model._
-import seqexec.web.client.model.Pages._
 import seqexec.web.client.actions._
-import scala.concurrent.Future
-import scala.scalajs.concurrent.JSExecutionContext.Implicits.queue
 
 class NavigationHandler[M](modelRW: ModelRW[M, Pages.SeqexecPages]) extends ActionHandler(modelRW) with Handlers {
   def handleNavigateTo: PartialFunction[Any, ActionResult[M]] = {
@@ -19,21 +16,7 @@ class NavigationHandler[M](modelRW: ModelRW[M, Pages.SeqexecPages]) extends Acti
 
   def handleSilentTo: PartialFunction[Any, ActionResult[M]] = {
     case NavigateSilentTo(page) =>
-      val effect = page match {
-        case SequencePage(i, id, _)          =>
-          Effect(Future(SelectIdToDisplay(i, id)))
-        case SequenceConfigPage(i, id, step) =>
-          Effect(Future(ShowStepConfig(i, id, step)))
-        case PreviewPage(i, id, step) =>
-          Effect(Future(SelectSequencePreview(i, id, step)))
-        case PreviewConfigPage(i, id, step) =>
-          Effect(Future(ShowPreviewStepConfig(i, id, step)))
-        case EmptyPreviewPage =>
-          Effect(Future(SelectEmptyPreview))
-        case _                               =>
-          VoidEffect
-      }
-      updatedSilent(page, effect)
+      updatedSilent(page)
   }
 
   def handle: PartialFunction[Any, ActionResult[M]] =
